UI.T0028 releng minor updates for windows ui

+++ plugins/com.codesourcery.installer.win32.x86/src/com/codesourcery/installer/os/Win32ActionsProvider.java
@@ -4,7 +4,7 @@
  *  are made available under the terms of the Eclipse Public License v1.0
  *  which accompanies this distribution, and is available at
  *  http://www.eclipse.org/legal/epl-v10.html
- * 
+ *
  *  Contributors:
  *     Mentor Graphics - initial API and implementation
  *******************************************************************************/
@@ -20,14 +20,14 @@ public class Win32ActionsProvider extends AbstractInstallPlatformActions {
 	@Override
 	public void bringToFront(Shell shell) {
 		/**
-		 * SWT doesn't provide a direct means to bring the shell to the front on 
-		 * all Windows versions.  <code>Shell.forceActive</code> will not on 
-		 * more modern versions of Windows.  This method uses the workaround 
+		 * SWT doesn't provide a direct means to bring the shell to the front on
+		 * all Windows versions.  <code>Shell.forceActive</code> will not on
+		 * more modern versions of Windows.  This method uses the workaround
 		 * given in:
 		 * https://bugs.eclipse.org/bugs/show_bug.cgi?id=192036
 		 */
-		int hFrom = OS.GetForegroundWindow();
-			
+		long hFrom = OS.GetForegroundWindow();
+
 		if (hFrom <= 0) {
 			OS.SetForegroundWindow(shell.handle);
 			return;
@@ -36,23 +36,23 @@ public class Win32ActionsProvider extends AbstractInstallPlatformActions {
 			return;
 		}
 		int pid = OS.GetWindowThreadProcessId(hFrom, null);
-	    int _threadid = OS.GetWindowThreadProcessId(shell.handle, null);
-	
-	    if (_threadid == pid) {
-	      OS.SetForegroundWindow(shell.handle);
-	      return;
-	    }
-	
-	    if (pid > 0) {
-	      if ( !OS.AttachThreadInput(_threadid, pid, true)) {
-	        return;
-	      }
-	      OS.SetForegroundWindow(shell.handle);
-	      OS.AttachThreadInput(_threadid, pid, false);
-	    }
-	
-	    OS.BringWindowToTop(shell.handle);
-	    OS.UpdateWindow(shell.handle);
-	    OS.SetActiveWindow(shell.handle);
+		int _threadid = OS.GetWindowThreadProcessId(shell.handle, null);
+
+		if (_threadid == pid) {
+			OS.SetForegroundWindow(shell.handle);
+			return;
+		}
+
+		if (pid > 0) {
+			if ( !OS.AttachThreadInput(_threadid, pid, true)) {
+				return;
+			}
+			OS.SetForegroundWindow(shell.handle);
+			OS.AttachThreadInput(_threadid, pid, false);
+		}
+
+		OS.BringWindowToTop(shell.handle);
+		OS.UpdateWindow(shell.handle);
+		OS.SetActiveWindow(shell.handle);
 	}
 }
